% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_excel}
\alias{download_excel}
\title{Download Excel File}
\usage{
download_excel(url, filepath = NULL, ext = ".xlsx", verbose = FALSE)
}
\arguments{
\item{url}{character URL}

\item{filepath}{character path file and extension for the downloaded file.
Leave as \code{NULL} to save as temporary file}

\item{ext}{character extension (ignored if \code{filepath} defined), must
be one of:
  \itemize{
    \item{.xlsx}
    \item{.xls}
  }}

\item{verbose}{logical print messages to console}
}
\value{
character name of temporary file
}
\description{
Given a URL download an excel spreadsheet to a temporary file
and test that the response is free of errors and that the correct file
type was found
}
\examples{
\dontrun{
  download_excel(url)
}
}
