% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\alias{download_excel}
\title{Download Files}
\usage{
download_file(
  url,
  filepath = NULL,
  ext = c(".xlsx", ".xls", ".zip"),
  files = NULL,
  verbose = FALSE
)

download_excel(
  url,
  filepath = NULL,
  ext = c(".xlsx", ".xls", ".zip"),
  files = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{character URL}

\item{filepath}{character path file and extension for the downloaded file.
Leave as \code{NULL} to save as temporary file}

\item{ext}{character extension, must be one of:
\itemize{
  \item{.xlsx}
  \item{.xls}
  \item{.zip}
}}

\item{files}{character vector of specific files to extract from a .zip}

\item{verbose}{logical print messages to console}
}
\value{
character name of temporary file/s
}
\description{
Given a URL download to a temporary file
and test that the response is free of errors and that the correct file
type was found. A .zip file will be automatically extracted after download
}
\examples{
\dontrun{
  download_file(url)
}
\dontrun{
  download_excel(url)
}
}
