% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_updates.R
\name{get_latest_download_url}
\alias{get_latest_download_url}
\title{Get URL for latest Dataset}
\usage{
get_latest_download_url(url, search_term)
}
\arguments{
\item{url}{character URL or web link}

\item{search_term}{character to match new data set link}
}
\value{
list
}
\description{
Get URL for latest Dataset
}
\examples{
\dontrun{
  dlpr::get_url_vic_median_rents_qrt(url, search_term)
}

}
