% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_median_rents_qrt.R
\name{get_url_vic_median_rents_qrt}
\alias{get_url_vic_median_rents_qrt}
\title{Get URL for latest Dataset}
\usage{
get_url_vic_median_rents_qrt()
}
\value{
list
}
\description{
Get URL for latest Dataset
}
\examples{
\dontrun{
  get_url_vic_median_rents_qrt()
}

}
