% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_payroll.R
\name{read_abs_payroll}
\alias{read_abs_payroll}
\title{Get ABS Weekly Payroll data}
\source{
https://www.abs.gov.au/statistics/labour/earnings-and-work-hours/weekly-payroll-jobs-and-wages-australia/latest-release
}
\usage{
read_abs_payroll(path = tempdir(), delete = TRUE)
}
\description{
Get ABS Weekly Payroll data
}
