% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aihw_mental_health.R
\name{read_aihw_care_episodes}
\alias{read_aihw_care_episodes}
\title{Australian Institute of Health and Welfare - Mental Health Data}
\usage{
read_aihw_care_episodes(
  url = urls$read_aihw_care_episodes,
  filename = tempfile()
)
}
\arguments{
\item{url}{Base url which contains links to all data, default url is: \link[https://labourmarketinsights.gov.au/our-research/internet-vacancy-index/]}

\item{filename}{Default filename is tempfile()}
}
\value{
data.frame
}
\description{
Extracting and cleaning yearly mental health data metrics
published by AIHW. The functions collect data for public healthcare episodes,
presentations and emergency departments, community care contacts and
prescriptions.The data is transformed into 'per 10k population' measures.
}
\examples{
\dontrun{
  df <- read_aihw_care_episodes()
  df <- read_aihw_emergency_presentations()
  df <- read_aihw_community_care()
  df <- read_aihw_prescriptions()
  }
}
