% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dese-childhood-services.R
\name{read_dese_childhood_services}
\alias{read_dese_childhood_services}
\title{DESE Childhood Services data}
\usage{
read_dese_childhood_services(
  url = urls$read_dese_childhood_services,
  filename = tempfile()
)
}
\arguments{
\item{url}{Base url which contains links to all data, default value is: \link[https://www.dese.gov.au/child-care-package/early-childhood-data-and-reports/quarterly-reports/]}

\item{filename}{Default file name is tempfile()}
}
\value{
data.frame
}
\description{
Reading childcare utilisation data from the DESE child care report/
}
\examples{
\dontrun{
  df <- read_dese_childhood_services()
}

}
