% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dss_payments.R
\name{read_dss_payments}
\alias{read_dss_payments}
\title{DSS Payment Demographic Data}
\usage{
read_dss_payments(url = urls$read_dss_payments)
}
\arguments{
\item{url}{Base url which contains links to all data, default value is: \link[https://data.gov.au/data/dataset/dss-payment-demographic-data/]}
}
\value{
data.frame
}
\description{
Extracting and cleaning payment data from the Department of
Social Services for each quarter back to September 2013. Data is transformed
into long, and is filtered by LGA and payment type.
}
\examples{
\dontrun{
  df <- read_dss_payments()
}

}
