% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_mobility.R
\name{read_google_mobility}
\alias{read_google_mobility}
\title{Google COVID Mobility Data}
\source{
\link{https://www.google.com/covid19/mobility/}
}
\usage{
read_google_mobility(
  region = "AU",
  sub_region_1 = "Victoria",
  filepath = NULL,
  attribution = FALSE
)
}
\arguments{
\item{region}{character code for Country. Defaults to "AU"}

\item{sub_region_1}{character code for State Defaults to "Victoria"}

\item{filepath}{character filename and path}

\item{attribution}{logical include Google attribution statement}
}
\value{
data.frame
}
\description{
download and transpose to long-format
"Google COVID-19 Community Mobility Reports".
}
\examples{
\dontrun{
  read_google_mobility()
}
}
