% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_internet_vacancy_report.R
\name{read_internet_vacancy}
\alias{read_internet_vacancy}
\title{National Skills Commission's Internet Vacancy Report}
\usage{
read_internet_vacancy(
  url = urls$read_internet_vacancy_report,
  dataset = c("skills", "occupations", "occupation_skills", "regions"),
  filename = tempfile()
)
}
\arguments{
\item{url}{Base url which contains links to all data, default url is: \link[https://www.jobsandskills.gov.au/work/internet-vacancy-index]}

\item{dataset}{character id of dataset to be downloaded. Must be one of:
\itemize{
  \item{skills}{anzsco_skill_level_states_and_territories}
  \item{occupations}{anzsco4_occupations_states_and_territories}
  \item{occupations_skills}{anzsco2_occupations_skill_level_states_and_territories}
  \item{regions}{anzsco2_occupations_ivi_regions}
}}

\item{filename}{Default filename is tempfile()}
}
\value{
data.frame
}
\description{
Extracting and cleaning monthly online job advertisements
 from JSA's Internet Vacancy Index Report.
}
\examples{
\dontrun{
  df <- read_read_internet_vacancy(dataset = 'skills')
}

}
