% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ivi.R
\name{read_ivi}
\alias{read_ivi}
\title{Get LMIP Internet Vacancy Index data}
\source{
https://lmip.gov.au/default.aspx?LMIP/GainInsights/VacancyReport
}
\usage{
read_ivi(path = tempdir(), delete = TRUE)
}
\description{
Get LMIP Internet Vacancy Index data
}
