% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jobseeker_payment_recipients.R
\name{read_jobseeker_payments}
\alias{read_jobseeker_payments}
\title{DSS JobSeeker Data}
\usage{
read_jobseeker_payments(url = urls$read_jobseeker_payments)
}
\arguments{
\item{url}{Base url which contains links to all data, default value is:
\link[https://data.gov.au/data/dataset/jobseeker-payment-and-youth-allowance-recipients-monthly-profile/]}
}
\value{
data.frame
}
\description{
Extracting and cleaning monthly JobSeeker payment data
from the Department of Social Services back to April 2020. Data
is transformed into long, and is filtered by SA2.
}
\examples{
df <- read_jobseeker_payments()
}
