% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_small_area_labour_markets.R
\name{read_salm}
\alias{read_salm}
\title{National Skills Commission's Small Area Labour Markets Estimates}
\usage{
read_salm(url = urls$read_salm)
}
\arguments{
\item{url}{Base url which contains links to all data, default value is:
\link[https://labourmarketinsights.gov.au/regions/small-area-labour-markets/]}
}
\value{
data.frame
}
\description{
Extracting and cleaning quarterly unemployment rates
by SA2 from the NSC's SALM Dataset.
}
\examples{
\dontrun{
  df <- read_salm()
}
}
