% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_cime_stats.R
\name{read_vic_crime_stats}
\alias{read_vic_crime_stats}
\title{Victorian Crime Statostics, Counts and rates.}
\source{
\url{https://www.crimestatistics.vic.gov.au/}
}
\usage{
read_vic_crime_stats(
  dataset = NULL,
  lga_input = NULL,
  suburb_input = NULL,
  dim_code_input = NULL,
  offence_div_code_input = NULL
)
}
\arguments{
\item{dataset}{Required. select the database c("offence_by_location_by_lga", "offence_by_postcode", "offence_rate_by_lga")}

\item{lga_input}{Optional. Filter by LGA (i.e. "Alpine")}

\item{suburb_input}{Optional. Filter by Suburb (i.e. "Dederang")}

\item{dim_code_input}{Optional. Filter by dim code (i.e. 11)}

\item{offence_div_code_input}{Optional. Filter by Offence Code (i.e. "A")}
}
\value{
data.frame
}
\description{
Each quarter the Victorian
\link[https://www.crimestatistics.vic.gov.au/]{Crime Statistics Agence (CSA)}
publish data on the number of criminal offenses and related information. Data is included for
each Local Government Area
}
\note{
Replaces the data import component of
\itemize{
  \item \url{https://github.com/djpr-data/osd-nous-dashboard/blob/main/processing/offences-by-location-type-lga.R}
  \item \url{https://github.com/djpr-data/osd-nous-dashboard/blob/main/processing/offence-by-division-postcode.R}
  \item \url{https://github.com/djpr-data/osd-nous-dashboard/blob/main/processing/offences-rate-by-division-lga.R}
}
}
\examples{
\dontrun{
  lga_offence_rates <- read_vic_crime_stats(dataset = "offence_rate_by_lga")

  Alpine_lga_offence_rates <- read_vic_crime_stats(dataset = "offence_rate_by_lga", lga_input = "Alpine")
}
}
