% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_median_rents_qrt.R
\name{read_vic_median_rents}
\alias{read_vic_median_rents}
\title{Victorian Rental Properties, Count and Median prices.}
\format{
A data frame with 11 variables. This is an active dataset rows increase weekly:
\describe{
  \item{district \code{character}}{Broad area districts:
    \itemize{
          \item{Barwon South West}
          \item{Grampians}
          \item{Loddon Mallee}
          \item{Hume}
          \item{Gippsland}
          \item{North and West Metro}
          \item{Eastern Metro}
          \item{Southern Metro}
          \item{METRO NON-METRO}
    }
  }
  \item{area \code{character}}{Local Government Area (LGA)}
  \item{property \code{character}}{Property type based on bedroom number}
  \item{date \code{date}}{Record date fields converted from "Year Month" format}
  \item{variable \code{character}}{Type of data}
  \item{value \code{numeric}}{Value}
  \item{mon \code{character}}{Month}
  \item{mon_num \code{character}}{Month number}
  \item{yr \code{numeric}}{Year}
}
}
\source{
\url{https://www.dffh.vic.gov.au/publications/rental-report}
}
\usage{
read_vic_median_rents(
  url = urls$read_vic_median_rents,
  include = c("all", "lga", "metro"),
  test = FALSE
)
}
\arguments{
\item{url}{character filename and path}

\item{include}{character the data set includes various spatial groupings
that can be automatically filtered. Options are:
\itemize{
  \item{all}
  \item{lga}
  \item{metro}
}}

\item{test}{logical for testing to avoid repeatedly downloading data}
}
\value{
tibble
}
\description{
Each quarter the Victorian
\link[https://www.dffh.vic.gov.au/]{Department of Families Fairness and Housing (DFFH)}
publish data on the number of rental properties and the median price. Data is included for
each Local Government Area
}
\note{
Replaces the data import component of \url{https://github.com/djpr-data/osd-nous-dashboard/blob/main/processing/rents-by-lga.R}
}
\examples{
\dontrun{
  rents_all <- read_vic_median_rents_qrt()

  rents_lga <- read_vic_median_rents_qrt(include = "lga")
}
}
