% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_median_rents_qrt.R
\name{read_vic_median_rents_qrt}
\alias{read_vic_median_rents_qrt}
\title{Get median rental prices for Victorian housing}
\source{
https://www.dffh.vic.gov.au/publications/rental-report
}
\usage{
read_vic_median_rents_qrt()
}
\value{

}
\description{
Get median rental prices for Victorian housing
}
