% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_property_prices.R
\name{read_vic_property_prices}
\alias{read_vic_property_prices}
\title{Victorian Property prices, yearly and recent median prices.}
\format{
A data frame with 11 variables. This is an active dataset rows increase weekly:
\describe{
  \item{locality \code{factor}}{locality or spatial grouping of data}
  \item{cdate \code{character}}{date as shown in spreadsheets}
  \item{date \code{date}}{last day in the period indicated by \code{time_scale}}. Therefore,
  last day of the year or last day of the quarter.
  \item{median \code{numeric}}{median dollar value of property sales within the \code{time_scale}}
  \item{time_scale \code{factor}}{indicating time period of the record:
    \itemize{
      \item{quarter}
      \item{year}
    }
  }
  \item{type \code{factor}}{properties are one of:
    \itemize{
      \item{house}
      \item{unit}
      \item{vacant}
    }}
}
}
\source{
\url{https://www.land.vic.gov.au/valuations/resources-and-reports/property-sales-statistics}
}
\usage{
read_vic_property_prices(
  url = urls$read_vic_property_prices,
  include = c("all", "year", "quarterly"),
  test = FALSE
)
}
\arguments{
\item{include}{character the data set includes multiple temporal groupings
that can be automatically filtered. Options are:
\itemize{
  \item{all}
  \item{year}
  \item{quarterly}
}}

\item{test}{logical for testing to avoid repeatedly downloading data}
}
\value{
tibble
}
\description{
The Victorian Property Sales Report is released quarterly in
March, June, September and December. It covers median sale prices by suburb/locality
for houses, units and vacant land over a period of 15 months.
}
\note{
Replaces the data import component of \url{https://github.com/djpr-data/osd-nous-dashboard/blob/main/processing/house-and-land-prices.R}
}
\examples{
\dontrun{
  property_prices_all <- read_vic_property_prices()

  rents_yearly <- read_vic_property_prices(include = "year")
}
}
