% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vic_schools_fte.R
\name{read_vic_schools_fte}
\alias{read_vic_schools_fte}
\title{School Enrolments for Victoria}
\usage{
read_vic_schools_fte()
}
\value{

}
\description{
School Enrolments for Victoria
}
\examples{
\dontrun{
  read_vic_schools_fte()
}
}
